requirements.txt
keyboard>=0.13.5
pystray>=0.19.4
Pillow>=9.0.0
pyperclip>=1.8.2
pyinstaller>=5.13.0